Move mqtt protocol utils into separate package
