+++ py_hpsu_monitor/workers/elster_register_mqtt_logger.py
@@ -1,6 +1,5 @@
 # pyright: reportUnnecessaryIsInstance=warning
-import json
-from typing import Any, Dict, List
+from typing import List
 
 import asyncio_mqtt
 
@@ -8,10 +7,13 @@ from ..config import MqttConfig
 from ..elster_protocol.elster_frame import ElsterFrame, ElsterReadResponseFrame
 from ..elster_protocol.register_definitions import group_register_definitions_by_index
 from ..elster_protocol.register_types import (
-    NumberRegisterDefinition,
     RegisterDefinition,
-    RegisterValue,
 )
+from ..mqtt_protocol.configuration_topic import (
+    get_configuration_payload,
+    get_configuration_topic,
+)
+from ..mqtt_protocol.state_topic import get_state_payload, get_state_topic
 from ..utils.publish_subscribe_topic import PublishSubscribeTopic
 
 
@@ -49,99 +51,3 @@ async def mqtt_log_elster_registers(
             await mqtt_client.publish(
                 topic=state_topic, payload=state_payload, retain=True
             )
-
-
-def get_configuration_payload(
-    mqtt_config: MqttConfig, register_definition: RegisterDefinition
-):
-    sensor_name = get_sensor_name(mqtt_config, register_definition)
-    state_topic = get_state_topic(mqtt_config, register_definition)
-    return json.dumps(
-        {
-            **{
-                "name": sensor_name,
-                "state_topic": state_topic,
-                "value_template": "{{ value_json.value }}",
-                "device": {
-                    "identifiers": [mqtt_config.device.id],
-                    "manufacturer": mqtt_config.device.manufacturer,
-                    "model": mqtt_config.device.model,
-                    "name": mqtt_config.device.name,
-                },
-                "unique_id": sensor_name,
-            },
-            **get_device_class(register_definition),
-        }
-    )
-
-
-def get_state_payload(register_value: RegisterValue[Any]):
-    return json.dumps(
-        {"timestamp": register_value.timestamp, "value": register_value.value}
-    )
-
-
-def get_configuration_topic(
-    mqtt_config: MqttConfig, register_definition: RegisterDefinition
-):
-    return mqtt_config.configuration_topic_template.format(
-        device_id=get_sensor_name(mqtt_config, register_definition)
-    )
-
-
-def get_state_topic(mqtt_config: MqttConfig, register_definition: RegisterDefinition):
-    return mqtt_config.state_topic_template.format(
-        device_id=get_sensor_name(mqtt_config, register_definition)
-    )
-
-
-def get_sensor_name(mqtt_config: MqttConfig, register_definition: RegisterDefinition):
-    return f"{mqtt_config.device.id}-{register_definition.name}"
-
-
-def get_device_class(register_definition: RegisterDefinition) -> Dict[str, Any]:
-    if (
-        isinstance(register_definition, NumberRegisterDefinition)
-        and register_definition.unit
-    ):
-        device_class = device_class_by_unit.get(register_definition.unit, None)
-        state_class = state_class_by_unit.get(register_definition.unit, None)
-
-        return {
-            "unit_of_measurement": register_definition.unit,
-            **(
-                {
-                    "device_class": device_class,
-                }
-                if device_class
-                else {}
-            ),
-            **(
-                {
-                    "state_class": state_class,
-                }
-                if state_class
-                else {}
-            ),
-        }
-
-    return {}
-
-
-device_class_by_unit = {
-    "°C": "temperature",
-    "W": "power",
-    "kW": "power",
-    "Wh": "energy",
-    "kWh": "energy",
-    "mbar": "pressure",
-}
-
-state_class_by_unit = {
-    "°C": "measurement",
-    "W": "measurement",
-    "kW": "measurement",
-    "Wh": "total_increasing",
-    "kWh": "total_increasing",
-    "mbar": "measurement",
-}
